requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
scikit-learn==1.3.2
statsmodels==0.14.0
streamlit==1.28.2
gunicorn==21.2.0
slack-sdk==3.25.0
Pillow==10.1.0
schedule==1.2.0
kaleido==0.2.1 